Work in progress towards supporting cross-compiling with sysroots
The goal is to treat the scratchbox arm target installation as a sysroot for
cross-compilation, together with the Qt SDK's arm cross compiler.

This eliminates funny issues such as qmake/moc/uic being arm binaries
and therefore horribly slow.

There are three main obstacles that I've encountered so far:

    1) The scratchbox target may contain silly absolute symlinks, such
as /usr/lib/libglib-2.0.so being a symlink to /lib/libglib-2.0.so instead
of being a relative symlink to ../lib/liblib-2.0.so. Absolute symlinks
for libraries don't play well with sysroots because they end up pointing
outside into the host. A separate shell script addresses this issue by
modifying the sysroot.

    2) Qt's configure isn't perfect in handling sysroots, we need to help it
to understand the target architecture and endianess.

     3) The compiler name is unfortunately not a variable but hardcoded in the
make spec, of which in turn there exists only one arm cross compilation one.
Unfortunately the harmattan toolchain shipped with the Qt SDK does not comply
to the naming, therefore we're going to need another helper script to fix it up
for us. In the meantime the setup in common.sh tries to verify that the environment
is sane, i.e. the compiler is available.